DataIngestionFramework/requirements.txt
azure-kusto-data==4.6.3
azure-eventhub==5.13.0
tenacity==9.0.0